+++ deno.json
@@ -1,0 +1,20 @@
+{
+  "imports": {
+    "@types/express": "npm:@types/express@^5.0.1",
+    "@types/express-session": "npm:@types/express-session@^1.18.1",
+    "connect-mongo": "npm:connect-mongo@^5.1.0",
+    "express": "npm:express@^5.1.0",
+    "express-session": "npm:express-session@^1.18.1",
+    "mongodb": "npm:mongodb@^6.15.0"
+  },
+  "compilerOptions": {
+    "jsx": "react-jsx",
+    "jsxImportSource": "preact"
+  },
+  "tasks": {
+    "seed-database":"deno run -A src/scripts/seedDatabase.ts",
+    "dev": "deno run --watch-hmr -A src/index.ts",
+    "start": "deno run -A src/index.ts"
+  },
+  "vendor": true
+}
